+++ doc/whats_new.rst
@@ -31,7 +31,9 @@ Changelog
 
     - Add example of time-frequency decoding with CSP by `Laura Gwilliams`_
 
-    - Add :class:`mne.decoding.SPoC` to fit and apply spatial filters based on continuous target variable, by `Jean-Remi King`_ and `Alexandre Barachant`_
+    - Add :class:`mne.decoding.SPoC` to fit and apply spatial filters based on continuous target variables, by `Jean-Remi King`_ and `Alexandre Barachant`_
+
+    - Add Fieldtrip's electromyogram dataset, by `Alexandre Barachant`_
 
 BUG
 ~~~